Gráfico event manager incluyendo días con número de eventos = 0

+++ hermesManager/model/src/main/java/es/udc/lbd/hermes/model/events/ListaEventosYdias.java
@@ -1,6 +1,7 @@
 package es.udc.lbd.hermes.model.events;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.util.List;
 
 public class ListaEventosYdias implements Serializable{
@@ -8,11 +9,11 @@ public class ListaEventosYdias implements Serializable{
 	private static final long serialVersionUID = 1L;
 	
 	List<String> fechas;
-	List<Long> nEventos;
+	List<BigInteger> nEventos;
 
 	
 	public ListaEventosYdias(){}
-	public ListaEventosYdias(List<String> fechas, List<Long> nEventos) {
+	public ListaEventosYdias(List<String> fechas, List<BigInteger> nEventos) {
 		super();
 		this.fechas = fechas;
 		this.nEventos =nEventos;
@@ -26,11 +27,11 @@ public class ListaEventosYdias implements Serializable{
 		this.fechas = fechas;
 	}
 	
-	public List<Long> getnEventos() {
+	public List<BigInteger> getnEventos() {
 		return nEventos;
 	}
 	
-	public void setnEventos(List<Long> nEventos) {
+	public void setnEventos(List<BigInteger> nEventos) {
 		this.nEventos = nEventos;
 	}
 }
